refactor: rename member-status to membership-status
